requirements.txt
asgiref==3.5.2
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2.binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.13.0